+++ proto-build/src/main.rs
@@ -21,8 +21,8 @@ const PROTO_DIR: &str = "../proto/src/prost/";
 const CHAIN_MAIN_REV: &str = "v3.3.3";
 const CHAIN_MAIN_DIR: &str = "../third_party/chain-main";
 
-const TERRA_REV: &str = "v0.5.18";
-const TERRA_DIR: &str = "../third_party/terra";
+const LUNA_CLASSIC_REV: &str = "v0.5.18";
+const LUNA_CLASSIC_DIR: &str = "../third_party/luna_classic";
 
 /// A temporary directory for proto building
 const TMP_BUILD_DIR: &str = "/tmp/tmp-protobuf/";
@@ -73,7 +73,7 @@ fn main() {
     update_submodules();
     output_commit_versions(&tmp_build_dir);
     compile_chain_main_protos_and_services(&tmp_build_dir);
-    compile_terra_protos_and_services(&tmp_build_dir);
+    compile_luna_classic_protos_and_services(&tmp_build_dir);
     copy_generated_files(&tmp_build_dir, &proto_dir);
 }
 
@@ -122,25 +122,25 @@ fn update_submodules() {
     run_git(&["-C", CHAIN_MAIN_DIR, "fetch"]);
     run_git(&["-C", CHAIN_MAIN_DIR, "reset", "--hard", CHAIN_MAIN_REV]);
 
-    info!("Updating terra submodule...");
+    info!("Updating luna_classic submodule...");
     run_git(&[
         "-C",
-        TERRA_DIR,
+        LUNA_CLASSIC_DIR,
         "submodule",
         "update",
         "--init",
         "--recursive",
     ]);
-    run_git(&["-C", TERRA_DIR, "fetch"]);
-    run_git(&["-C", TERRA_DIR, "reset", "--hard", TERRA_REV]);
+    run_git(&["-C", LUNA_CLASSIC_DIR, "fetch"]);
+    run_git(&["-C", LUNA_CLASSIC_DIR, "reset", "--hard", LUNA_CLASSIC_REV]);
 }
 
 fn output_commit_versions(out_dir: &Path) {
     let path = out_dir.join("CHAIN_MAIN_COMMIT");
     fs::write(path, CHAIN_MAIN_REV).unwrap();
 
-    let path = out_dir.join("TERRA_COMMIT");
-    fs::write(path, TERRA_REV).unwrap();
+    let path = out_dir.join("LUNA_CLASSIC_COMMIT");
+    fs::write(path, LUNA_CLASSIC_REV).unwrap();
 }
 
 fn compile_chain_main_protos_and_services(out_dir: &Path) {
@@ -196,14 +196,14 @@ fn compile_chain_main_protos_and_services(out_dir: &Path) {
     info!("=> Done!");
 }
 
-fn compile_terra_protos_and_services(out_dir: &Path) {
+fn compile_luna_classic_protos_and_services(out_dir: &Path) {
     info!(
-        "Compiling terra .proto files to Rust into '{}'...",
+        "Compiling luna_classic .proto files to Rust into '{}'...",
         out_dir.display()
     );
 
     let root = env!("CARGO_MANIFEST_DIR");
-    let sdk_dir = Path::new(TERRA_DIR);
+    let sdk_dir = Path::new(LUNA_CLASSIC_DIR);
 
     let proto_includes_paths = [
         format!("{}/../proto", root),