+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 ### Added
 - Add polling interval argument or function for setting the polling interval on event filters and pending transactions.
 - Add `CosmosParser` to support parsing both Protobuf and Proto3 JSON mapping of standard Cosmos and `crypto.org` messages.
-- Add Terra special messages to `proto-build`.
+- Add Luna Classic special messages to `proto-build`.
 ### Removed
 - Delete duplicate C-Sytle functions for Cosmos SDK in wasm binding.